KUDU-1315. Add simple script to build a source release
This adds a new python script meant for interactive use for building
an Apache-style source release.

The source release tarball is simply the output of 'git archive'. The
purpose of the script is to perform some pre-flight checks to make sure
the release seems sane, and to help the user gpg-sign the release upon
completion. It's not fancy, but should ensure that the release process
is repeatable (two committers running this script from the same revision
should result in 100% identical release artifacts).

In order to generate the artifact name, I moved the version string into
a text file at the repository root instead of embedding it into the
CMakeLists.txt file.

Change-Id: Ie5753c4cc822dfbc86210356850ea1b76ab8e7a2
Reviewed-on: http://gerrit.cloudera.org:8080/1930
Reviewed-by: Jean-Daniel Cryans
Tested-by: Todd Lipcon <todd@apache.org>