build: properly clean up Python build artifacts
A change to the location of libkudu_client.so is percolating through
gerrit and has exposed a problem in the Python build: Cython shared objects
are not being rebuilt as often as they should be. The result is that at
runtime, these shared objects can't find libkudu_client.so because their
rpaths are stale.

Let's give them a hand with some explicit cleanup.

Change-Id: I31125cf7659fb37f88daabd1a35646526718ccd8
Reviewed-on: http://gerrit.cloudera.org:8080/1950
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins