Fix flakiness of client_failover-itest
The reason this test was flaky is that there is a race between
determining that all of the replicas in a config have replicated an opid
and that opid being committed. We were also missing the
--master_add_server_when_underreplicated=false flag on the test, causing
a race between the master and the test driver.

This patch exposes getting the committed OpId from Consensus as an RPC
(previously, only the last replicated opid was exposed). It also adds a
helper function for testing purposes to wait until a specific OpId has
been committed.

Looped 100x and they all passed:
http://dist-test.cloudera.org/job?job_id=mpercy.1454486819.10566

Change-Id: I56ea7523e212c1b0f67b1498ba5139e4c5537519
Reviewed-on: http://gerrit.cloudera.org:8080/1952
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>
(cherry picked from commit 1a24338ad60a8842d1ae5e227f8f03e58faea8c0)
Reviewed-on: http://gerrit.cloudera.org:8080/2032