tsan: remove libstdcxx race annotations
Now that we run with a TSAN-instrumented libstdcxx, we no longer need to
suppress various std::string and shared_ptr functions.

However, the version of libstdcxx that we run against has a few minor
data races that don't cause much trouble in practice, but do trip up
TSAN. So, this adds two patches to libstdcxx to fix the issues:

- the first is an upstream patch[1] which fixes std::string
- the second is a similar patch for tr1::shared_ptr which is still used in the
  client code. This patch isn't upstream, but basically just follows the example
  from the C++11 shared_ptr implementation.

With these fixes, TSAN seems to run cleanly.

[1] https://gcc.gnu.org/viewcvs/gcc?view=revision&revision=227403

Change-Id: Iada66d777c1a1aa348759b5697706e6bcd335945
Reviewed-on: http://gerrit.cloudera.org:8080/1983
Tested-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>