[python] - Allow to scan null values
Currently scanning a column with null values causes a crash when
the client tries to get the null value.

This adds an is_null() check so that we add 'None' to the tuple
if the value is null in the tablet.

This also alters the scanner test to produce and read null values.

Change-Id: I0264c30d1177b31c92db1669c6b1883d1ec691ba
Reviewed-on: http://gerrit.cloudera.org:8080/1989
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <david.alves@cloudera.com>