[python] - Allow to scan null values
Currently scanning a column with null values causes a crash when
the client tries to get the null value.

This adds an is_null() check so that we add 'None' to the tuple
if the value is null in the tablet.

This also alters the scanner test to produce and read null values.

Change-Id: If340213cf5daba50c60d271530c21b109c8cd2f3
Reviewed-on: http://gerrit.cloudera.org:8080/1989
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <david.alves@cloudera.com>
Reviewed-on: http://gerrit.cloudera.org:8080/2039
Reviewed-by: Jean-Daniel Cryans