[python] - Allow to set a scanner to be fault tolerant
Making a scanner fault tolerant is useful in any case, but for cases
where hash partitioning is not being used, fault tolerance has the side
effect of also making the scanner return ordered results. This is a very
useful feature for timeseries use cases as it avoids an additional sort.

This just makes it so that it is possible make a python scanner fault
tolerant by calling the adequate method on the c++ client.

Change-Id: Iaf27778cde442f1da66e42abe8bd0f06e715dcd9
Reviewed-on: http://gerrit.cloudera.org:8080/1993
Tested-by: David Ribeiro Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>
Reviewed-on: http://gerrit.cloudera.org:8080/2040
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans