Change timestamp printing to ISO-8601 (with microseconds)
Dan raised the concern in a recent review that we're not adhering to the
standard while printing timestamps. The main point of this was to mimic
impala's behavior but then we would also print 'GMT', which impala does
not, meaning we aren't really mimicking impala either.

This changes timestamp printing to adhere to ISO-8601, hopefully allowing
standard date parsers to more easily process the output, if needed.

Change-Id: I421e3595af2b21eee6ec22606b6046e470559ad4
Reviewed-on: http://gerrit.cloudera.org:8080/1995
Tested-by: David Ribeiro Alves <dralves@apache.org>
Reviewed-by: Todd Lipcon <todd@apache.org>