docs: update build-from-source instructions for blessed build layout
I'm not convinced that this makes sense; in my opinion, installation.adoc
isn't intended for Kudu developers, but for people running on platforms for
which we lack prebuilt binaries. JD pointed out that the Java client build
muddies this somewhat, as it's reasonable to expect these people to want to
build the Java client, and by default maven will run unit tests, which will
fail unless they're using the blessed build layout (or pass -DbinDir).

Anyway, the blessed build layout doesn't actually hurt casual
build-from-source people; it's just more complexity.

I also removed the various mentions of openssl-devel as with the dlopen()
change to squeasel it's no longer necessary for building.

Change-Id: Ic482c084397d13a0fd9e0b3e710449bc5cd866c4
Reviewed-on: http://gerrit.cloudera.org:8080/1997
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans