threadpool: prohibit shutdown and wait calls from pool member threads
If a thread pool thread calls a shutdown or wait function on its own pool,
it'll deadlock. This patch introduces some checks that will turn the
deadlock into a crash as well as a test to exercise it.

The solution isn't particularly elegant, though I think it's less overhead
than a TLS-based implementation.

Change-Id: I1e00b251e0b798648b29d75c5cd82a25307129b4
Reviewed-on: http://gerrit.cloudera.org:8080/2013
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>