[c++ client] - Expose the schema being scanned in KuduScanBatch
This exposes the schema being scanned in the KuduScanBatch and KuduScanBatch::RowPtr
allowing to pass pointer/references to these around when in use without an accompanying
schema, which was hard to infer in any case since the scanner projection is set with
just column names.

Note that this is just exposing the schema information that was already there, just
in KuduSchema form (vs. the internal Schema) which is more suited for user consumption.

Change-Id: I79dde3ece27c5122fba9c69c70ef9e25d641c1e1
Reviewed-on: http://gerrit.cloudera.org:8080/2020
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-on: http://gerrit.cloudera.org:8080/2037