KUDU-966. Construct and cache the master registration PB at startup
On some clusters, calls to determine the FQDN can run very slowly if the DNS is
not set up properly. This can cause GetMasterRegistration RPCs to take several
seconds, and cause the master to basically fall over.

This moves the call to construct the registration protobuf into the startup
sequence, so that the RPC can be satisfied by a simple copy from the
cached value.

No new tests since this is a performance improvement which is only visible
on machines with broken DNS.

Change-Id: I79bac224d905cad579a6f4fa6552b8ea17af6b4a
Reviewed-on: http://gerrit.cloudera.org:8080/2053
Reviewed-by: Jean-Daniel Cryans
Tested-by: Kudu Jenkins