[c++ client] - Remove KuduSchema from KuduScanBatch::RowPtr
Todd manifested concern that this was making the pointer less like a pure
pointer, i.e. more bloated.

The main reason this was made available here at all was that the python
client was still using KuduRowResult and thus needed it. Since that is
no longer the case this field (and its accessors) can be removed.

Change-Id: I8aff27ac92babf4dc8b917cbe01c04e23b15de2a
Reviewed-on: http://gerrit.cloudera.org:8080/2076
Reviewed-by: Jean-Daniel Cryans
Tested-by: Jean-Daniel Cryans