Integrate ColumnPredicate into client and server
This commit integrates the new ColumnPredicate type into the ScanSpec class. The
predicate encoder functionality has been subsumed into the ScanSpec and
ColumnPredicate classes.

In addition, we now 'lift' implicit predicates from the lower and upper primary
key bounds into the set of column predicates when the scan spec is optimized.

Scans using column predicates will fail to complete if the server does not
support the column predicate feature, so users will need to update the server
version when using a client version with this patch.

Finally, this commit refactors row_key-util into key_util, and makes the utility
methods able to work on primary keys as well as range keys.

Change-Id: Ife6852680b7f59fddee688e5702c1a70944f7622
Reviewed-on: http://gerrit.cloudera.org:8080/2138
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>