Update NOTICE.txt to include Apache copyright
According to http://apache.org/legal/src-headers.html#notice the
NOTICE.txt file should follow a particular template. This patch updates
the file to follow the template.

After going through all of the licenses in LICENSE.txt, I believe none
of them require special notification aside from inclusion in LICENSE.txt
for a source distribution.

I kept the mention of Cloudera as a courtesy, since Cloudera was the
original donor of Kudu to the ASF. This has been done in other Apache
projects as well.

Change-Id: I9bfb81f786e33c174eb625fd92b86dcffd97877b
Reviewed-on: http://gerrit.cloudera.org:8080/2108
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans