Improvements to download-thirdparty.sh for low bandwidth connections
When internet bandwidth is harder to come by than disk space, it's nice
to have an option to keep thirdparty archives around instead of
re-downloading them all the time. By setting the
NO_REMOVE_THIRDPARTY_ARCHIVES environment variable, the downloaded
source tarballs for thirdparty will not be removed and will be reused
the next time download-thirdparty.sh attempts to fetch the artifact. If
the artifact is corrupted, it is removed and the script attempts to
re-download it. This retry only happens once.

Change-Id: I1c7ca6994d2c0251676ebe1397cf4fccfb1d59f9
Reviewed-on: http://gerrit.cloudera.org:8080/2122
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins