Add ASL2.0 license headers to files missing it
Ran RAT manually and these were the files missing ASL that were not
licensed under some other license, such as BSD or MIT, and that were of
a file format that supports adding comments.

I'm working on integrating RAT but I didn't want to block these fixes or
the release on that.

Change-Id: I6ace596ce1fec648cc42edaed8fbca1d2bd941e0
Reviewed-on: http://gerrit.cloudera.org:8080/2178
Reviewed-by: Jean-Daniel Cryans
Tested-by: Kudu Jenkins