Add RAT excludes file and parsing script for releases
The RAT tool itself doesn't have very robust support for exclude files
(only filenames, not paths, are supported), so I've written a little
Python script to parse the output of the RAT XML report and apply rules
from a simple path-globbing exclude file. Instructions in the README.

Ultimately, it would be nice to fully automate downloading and running
RAT, but this is better than nothing.

Change-Id: Id6850e71665c0846b5f57d0527b930a629f0ab4a
Reviewed-on: http://gerrit.cloudera.org:8080/2180
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans