KUDU-1328: crash when remote bootstrap session races with tablet deletion
An innocuous log statement introduced in commit b841512 can cause a crash if
a remote bootstrap request for a particular tablet arrives as the tserver
deletes that tablet.

I'm still working on an integration test that faithfully reproduces the
crash, but until then, here's a low-risk fix for the log statement that can
go into 0.7.

Change-Id: I6a487ba4f37383479170519866509f3f3d88dc5d
Reviewed-on: http://gerrit.cloudera.org:8080/2193
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans