catalog_manager: fix a use-after-free on TabletInfo
This fixes an issue seen in a new stress test that Adar is working on:

T1:
- HandleAssignCreatingTablet calls CreateTabletInfo, which returns a raw
  pointer 'replacement'
- we added that raw pointer to the table's tablet map (which also stores it as raw)

T2:
- GetTableLocations() constructs a scoped_refptr to the TabletInfo object,
  acting as its first referer.
- GetTableLocations() finishes and drops the scoped_refptr, deleting the TabletInfo

T1:
- we try to dereference 'replacement', and crash, because T2 freed the pointer.

This patch fixes the issue by using a scoped_refptr to hold the 'replacement'
tablet before exposing it to other threads.

No new unit test here, because it was easily reproduced by the stress test that
Adar's working on, which will be committed soon.

Change-Id: I67259d54f03bded7c7bbbea677093f39e0efa528
Reviewed-on: http://gerrit.cloudera.org:8080/2213
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>