build: fix ignoring of clock sync errors in flaky test list
Previously, I tried to add this same functionality, but used 'grep' instead of
'zgrep'. In the production test slaves, the test logs are gzipped, so we need
to 'zgrep'. It turns out that zgrep is also smart enough to grep non-compressed
files, so it's safe to use it regardless of the file extension.

Change-Id: Ic8645f6f046700a28d3306e0db1be9dbfb06a433
Reviewed-on: http://gerrit.cloudera.org:8080/2218
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins