Move design docs to the new folder and change the extension
This moves the design docs that we had sprinkled all over the code base to the
new design doc folder and changes the extension to .md, as well as adds them
to the design doc index. Though this doesn't do any kind of thorough markdown
conversion it does introduce '```' where appropriate to avoid that our nice ascii
diagrams become garbled garbage when visualized on github.

Change-Id: Iebeaba9e975453458ac993131dec6074c59d0eeb
Reviewed-on: http://gerrit.cloudera.org:8080/2227
Reviewed-by: Dan Burkert <dan@cloudera.com>
Tested-by: David Ribeiro Alves <david.alves@cloudera.com>