rpc: call-level feature flags
This adds a new API RpcController::RequireServerFeature on the client side, and
ServiceIf::SupportsFeature on the service side. These APIs make it possible for
the client to require application-specific feature support on the server in
order for an individual RPC call to be serviced. If the feature is not
supported, the RPC is rejected and an appropriate error is returned.

This will be used in the future when adding features to existing RPC messages,
so that the server does not silently ignore additional fields.

These RPC features require a new RPC feature flag, APPLICATION_FEATURE_FLAGS, to
be supported. If the server does not support that RPC feature flag, then the
call will fail on the client side.

Change-Id: I46e46bcf80b93fc6cce50f37dd71a6e6539047f8
Reviewed-on: http://gerrit.cloudera.org:8080/2239
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>