Replace simpler instances of foo.Pass() with std::move(foo)
This commit was done primarily via this perl script:

  perl -p -i -e 's,([\w_]+)\.Pass\(\),std::move(\1),g' \
    $(find src -name \*.h -or -name \*.cc)

with one or two trivial manual fixups where the regex went wrong and
introduced a syntax error. Additionally, I reverted the changes to
gutil/callback* and gutil/bind* since those files are complicated
and seemed to cause a lot of trouble.

Change-Id: Ibff7624f2d21ee098dfba466ebfbb09fd2c9862d
Reviewed-on: http://gerrit.cloudera.org:8080/2246
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <dan@cloudera.com>