KUDU-1325: more crash avoidance during remote bootstrap and tablet deletion
Just as with KUDU-1328, here is yet another instance where remote bootstrap
doesn''t properly protect itself in the event of a race with tablet
deletion. KUDU-1337 is responsible for this sequence of events happening so
often, but it's worth addressing even if it were rare.

This time around, we'll deal with it by giving LogReader shared ownership.
Personally this feels like overkill (extra cognitive load when dealing with
LogReader just to prevent this race?) but it's the easiest to implement
given the design of the server-side part of remote bootstrap. I'm still
working on reproducing these races reliably in an integration test.

What else is interesting? Turns out std::make_shared() can't be used on
classes with private constructors. I've added a workaround in
ALLOW_MAKE_SHARED which is simple but entirely non-portable.

Change-Id: I6a01c9e6886bd8bf0e286fc3980babc512056286
Reviewed-on: http://gerrit.cloudera.org:8080/2265
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>