csd: always validate using Cloudera public maven repo
The schema-validator-maven-plugin artifact and its dependencies were missing
from Cloudera's public maven repo, which meant validation could only work
when the local maven repo was prepopulated with the plugin. This has since
been fixed, so it's now possible to validate the CSD from anywhere.

Due to that, I've also removed the "validation is optional" behavior: going
forward, building the CSD means validating it.

Change-Id: I120c874c219abf98e7c33737c34d248daaf8e5bb
Reviewed-on: http://gerrit.cloudera.org:8080/2311
Reviewed-by: Jean-Daniel Cryans
Tested-by: Kudu Jenkins