dist-test:  allow collecting the tmp dir of failed tests
When debugging certain types of failures, especially with randomized
tests, it can be very useful to grab the TEST_TMPDIR of any failed
runs, since that includes the WAL, data files, etc, of any minicluster
servers used in the test.

This adds a --collect-tmpdir flag to our dist_test wrapper which causes
the dist test task to create a tarball of its test tmpdir if the task
fails. It puts the tarball in the log directory which is considered the
"artifact" of the task, so that the 'fetch' command downloads them.

Change-Id: I57d42d93e2459bafebe433ade3a9aa7a45eaeabe
Reviewed-on: http://gerrit.cloudera.org:8080/2336
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>