KUDU-1341 (part 1): improve tablet_random-access-test to be a full integration test
This moves the randomized fuzz test part of tablet_random-access-test
into a new fuzz-itest integration test, and changes it to use a
MiniCluster instead of a LocalTabletWriter. This will allow us
to extend the test to also include tablet server restarts in the
next commit.

We use a MiniCluster instead of ExternalMiniCluster so the test can
still manually control flushes, compactions, etc.

While I was moving the code, I also updated the test cases to use
C++11 braced initializer syntax for the test case vectors.

Change-Id: I4a60436461facbeb39aa4ff5f7b2ba1c2b7a230c
Reviewed-on: http://gerrit.cloudera.org:8080/2346
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <david.alves@cloudera.com>