KUDU-1337. Avoid spurious remote bootstraps on DeleteTablet()
When a table is deleted, the master sends a DeleteTablet() RPC to every
replica of every tablet with the TABLET_DATA_DELETED parameter, which
indicates a "permanent" tablet deletion. If a follower services
DeleteTablet() before the leader does, it's possible for the leader to
react to the missing replica by asking the follower to remote bootstrap
itself.

This changes a bit of logic so that the TSTabletManager caches that a
tablet has been permamently deleted and refuses remote bootstrap
requests for a tablet if that tablet is in the deleted cache. This is a
transient cache and is reset upon process restart.

Change-Id: I852f8ac70e1f6496127598e5e02de5b72711ab2b
Reviewed-on: http://gerrit.cloudera.org:8080/2436
Tested-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>