whitepaper: remove use of the vldb.cls file
Per LEGAL-238, use of this file is in enough of a legal grey area that it's not
worth worrying about it. We don't really care about using the VLDB template
exactly (since this isn't a format VLDB submission), so this commit just
switches to using a built-in article document class.

I tweaked the margins and widths a bit so that the diagrams and tables still
render in a reasonable-looking way.

Change-Id: I5c743b93b249c6630c219769e606ddd109525815
Reviewed-on: http://gerrit.cloudera.org:8080/2373
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <david.alves@cloudera.com>