fuzz-itest: decrease update multipler in TSAN builds
I ran the fuzz test in TSAN 500 times on GCE and they all passed,
but we're still seeing write timeouts on EC2 slaves. Looking at
the test on my machine, the batches with 1000 updates against the
same row are very slow due to rolling through the deltas in
DeltaMemStore::CheckRowDeleted() on a TSAN build. On my machine,
it was taking ~15 seconds per batch, and if we got unlucky with the
random generation, we might have a few UPDATE ops in a row before
a flush, which could easily put it over the 60 second timeout,
especially on a slower box.

This just fixes the issue by not using such large batches on TSAN
builds.

Change-Id: If795cd0d01f9bc7cc2a05172d58c2fa7d7b2dbd7
Reviewed-on: http://gerrit.cloudera.org:8080/2401
Tested-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Todd Lipcon <todd@apache.org>