Disable SSL in curl
We don't use HTTPS in our curl utility calls, and removing it should
make it possible to push to the dist_test server (Ubuntu 14.04) from
CentOS 6, since they have different versions of libssl. We don't
distribute our own version of libssl because it's a security library
that should be patchable at the OS package level.

The only change is adding --without-ssl. The rest of the changes were
just reordering of the arguments to be alphabetical.

Change-Id: I7bec72cc0f382cd27897c3d84891db153019ed3e
Reviewed-on: http://gerrit.cloudera.org:8080/2434
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>