tablet: avoid opening CFile writers for empty flushes
I was looking at some failures of alter_table-randomized-test and
noticed that bootstrap is very slow in this test case. One of the
reasons is that the test triggers a lot of flushes of empty rowsets
as it replays alter tables. In one failed test run, there were 1395
instances of 'all input rows were GCed!'. These empty flushes were
wasting some IO to open cfile writers, write headers, and then delete
them.

This patch refactors some common code in the flush path between
AlterTable and normal flushes, and adds a short circuit for the case
of flushing an empty MRS.

Change-Id: I629f6a283f9963c0bb8e9be72d6aa4210f0aec72
Reviewed-on: http://gerrit.cloudera.org:8080/2441
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>