KUDU-1336. Add a simple preflight check for thirdparty build
We've had a lot of users hit common build problems like not using a new enough
C++11 compiler on RHEL6. This commit adds a simple set of pre-flight checks
that will hopefully make it easier for new users to build Kudu.

I tested this by running an ubuntu:14.04 docker image, running preflight.py and
installing whatever it asked me to until it passed. This seemed to cover all
of the packages mentioned in the install docs.

I also verified that this passed on a RHEL6 system with devtoolset.

Change-Id: I93ee4165bd560f9cd3f03877bd3011decc7e1a6f
Reviewed-on: http://gerrit.cloudera.org:8080/2450
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>