Move OverwriteWithPattern to new file optimized at -O3
This function ends up taking a fair amount of CPU in debug builds. We already
optimized it at -O3 for gcc, but clang doesn't support per-function
optimization. This commit just moves it to a new file and applies -O3 to that
file, so that clang builds also get an optimized implementation.

Hopefully this will improve flakiness of some tests that show up in ASAN/TSAN
builds due to very slow performance. This improved
'ts_recovery-itest --gtest_filter=\*969\*' from ~33s to ~25s on my laptop
in an ASAN build.

Change-Id: I1f18e6de3d5f5ab21b490ea4213c86b4d25706ba
Reviewed-on: http://gerrit.cloudera.org:8080/2478
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins