random: re-implement Random::Normal using std::normal_distribution
Since we were only using this one function from the WebRTC source code,
it was easier to just use std::normal_distribution than worry about
the proper way to attribute WebRTC (BSD license).

Tested with the existing unit test, and also modified it to dump a few
thousand values to a text file. I verified the resulting distribution
in R:

> sd(d$V1)
[1] 0.01002238
> mean(d$V1)
[1] 4.999974

(which is close enough given the sample size)

Change-Id: Id2101d1375618a87b370162c067f9bbf2035cf33
Reviewed-on: http://gerrit.cloudera.org:8080/2481
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>